requirements.txt
numpy
pandas
boto3
python-dotenv
pymongo
s3transfer
